Detect and correct for 1658B current quirk
The 1658B has one more digit of precision in its current values (per
page 11 of the programming manual). This means that the previous code
incorrectly displayed and set current values for that particular
variant.

This changeset detects the presence of a 1658B by examining the power
supply's maximum reported voltage. Though this is a little hacky, it
appears to be the only way to differentiate between the different models
in the 168xB line. If a 1685B is detected, a different current divider
is used.

It's worth noting that the GETD command (used to query the measured
current/voltage levels on the output) appears not to vary between
different PSUs in the line. However, I don't have an 1687B or 1688B with
which to test (yet) -- I'm just operating under the assumption that the
code as-written was working properly for them.